refactor: made this project more extendable and easier to maintain
