+++ docker/docker-compose.yml
@@ -1,0 +1,24 @@
+version: "3.8"
+
+services:
+  app:
+    build:
+      context: ..
+      dockerfile: docker/Dockerfile
+    image: llm-moe:latest
+    container_name: llm-moe-app
+    working_dir: /app
+    volumes:
+      - ..:/app
+      - ../checkpoints:/app/checkpoints
+      - ../logs:/app/logs
+      - ../outputs:/app/outputs
+    environment:
+      - TORCH_HOME=/app/checkpoints
+      - HF_HOME=/app/checkpoints
+      - MPLCONFIGDIR=/app/.mplconfig
+    # Requires Docker with NVIDIA Container Toolkit installed
+    gpus: all
+    ipc: host
+    shm_size: 16g
+    command: ["python", "src/main.py"]
