+++ src/model.py
@@ -1,0 +1,270 @@
+import math
+from typing import Optional, Tuple
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from torchtune.modules import RotaryPositionalEmbeddings
+
+from .config import ModelConfig, MoEModelConfig
+
+
+class Rotary(nn.Module):
+    def __init__(self, dim: int, max_seq_len: int):
+        super().__init__()
+        self.rope = RotaryPositionalEmbeddings(dim=dim, max_seq_len=max_seq_len, base=10000)
+
+    def forward(self, x_BTHD: torch.Tensor):
+        return self.rope(x_BTHD)
+
+
+class MultiHeadAttention(nn.Module):
+    def __init__(self, d_model: int, n_heads: int, max_seq_len: int, dropout: float = 0.1):
+        super().__init__()
+        self.d_model = d_model
+        self.n_heads = n_heads
+        self.d_k = d_model // n_heads
+
+        self.qkv = nn.Linear(d_model, d_model * 3, bias=False)
+        self.w_o = nn.Linear(d_model, d_model, bias=False)
+        self.rotary = Rotary(self.d_k, max_seq_len)
+        self.dropout = dropout
+
+    def forward(self, x):
+        batch_size, seq_len = x.size(0), x.size(1)
+
+        qkv = self.qkv(x).reshape(batch_size, seq_len, 3, self.n_heads, self.d_k)
+        qkv = qkv.permute(2, 0, 3, 1, 4)
+        Q, K, V = qkv[0], qkv[1], qkv[2]
+
+        Q = self.rotary(Q.transpose(1, 2)).transpose(1, 2)
+        K = self.rotary(K.transpose(1, 2)).transpose(1, 2)
+
+        attn_output = F.scaled_dot_product_attention(
+            Q, K, V, is_causal=True, dropout_p=self.dropout if self.training else 0.0
+        )
+        attn_output = attn_output.transpose(1, 2).reshape(batch_size, seq_len, self.d_model)
+        return self.w_o(attn_output)
+
+
+class FeedForward(nn.Module):
+    def __init__(self, d_model: int, d_ff: int, dropout: float = 0.1):
+        super().__init__()
+        self.linear1 = nn.Linear(d_model, d_ff, bias=False)
+        self.linear2 = nn.Linear(d_ff, d_model, bias=False)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x):
+        return self.linear2(self.dropout(F.silu(self.linear1(x))))
+
+
+class TransformerBlock(nn.Module):
+    def __init__(self, d_model: int, n_heads: int, d_ff: int, max_seq_len: int, dropout: float = 0.1):
+        super().__init__()
+        self.attention = MultiHeadAttention(d_model, n_heads, max_seq_len, dropout)
+        self.feed_forward = FeedForward(d_model, d_ff, dropout)
+        self.norm1 = nn.RMSNorm(d_model)
+        self.norm2 = nn.RMSNorm(d_model)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x):
+        attn_out = self.attention(self.norm1(x))
+        x = x + self.dropout(attn_out)
+        ff_out = self.feed_forward(self.norm2(x))
+        x = x + self.dropout(ff_out)
+        return x
+
+
+class Expert(nn.Module):
+    def __init__(self, d_model: int, d_ff: int, dropout: float = 0.1):
+        super().__init__()
+        self.linear1 = nn.Linear(d_model, d_ff, bias=False)
+        self.linear2 = nn.Linear(d_ff, d_model, bias=False)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x):
+        return self.linear2(self.dropout(F.silu(self.linear1(x))))
+
+
+class TopKRouter(nn.Module):
+    def __init__(self, d_model: int, num_experts: int, top_k: int = 2):
+        super().__init__()
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.gate = nn.Linear(d_model, num_experts, bias=False)
+        self.noise_std = 0.1
+
+    def forward(self, x: torch.Tensor) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
+        router_logits = self.gate(x)
+        if self.training and self.noise_std > 0:
+            noise = torch.randn_like(router_logits) * self.noise_std
+            router_logits = router_logits + noise
+        router_probs = F.softmax(router_logits, dim=-1)
+        top_k_logits, top_k_indices = torch.topk(router_logits, self.top_k, dim=-1)
+        top_k_weights = F.softmax(top_k_logits, dim=-1)
+        return top_k_weights, top_k_indices, router_probs
+
+
+class MixtureOfExperts(nn.Module):
+    def __init__(
+        self,
+        d_model: int,
+        d_ff: int,
+        num_experts: int = 8,
+        top_k: int = 2,
+        dropout: float = 0.1,
+        load_balancing_weight: float = 0.01,
+    ):
+        super().__init__()
+        self.num_experts = num_experts
+        self.top_k = top_k
+        self.load_balancing_weight = load_balancing_weight
+        self.experts = nn.ModuleList([Expert(d_model, d_ff, dropout) for _ in range(num_experts)])
+        self.router = TopKRouter(d_model, num_experts, top_k)
+
+    def forward(self, x: torch.Tensor) -> Tuple[torch.Tensor, Optional[torch.Tensor]]:
+        router_weights, expert_indices, router_probs = self.router(x)
+        output = torch.zeros_like(x)
+        for expert_idx in range(self.num_experts):
+            expert_mask = (expert_indices == expert_idx).any(dim=-1)
+            if expert_mask.any():
+                expert_input = x[expert_mask]
+                expert_output = self.experts[expert_idx](expert_input)
+                mask_for_expert = (expert_indices == expert_idx)
+                positions = mask_for_expert[expert_mask].float().argmax(dim=-1)
+                expert_weights = router_weights[expert_mask].gather(-1, positions.unsqueeze(-1)).squeeze(-1)
+                output[expert_mask] += expert_weights.unsqueeze(-1) * expert_output
+        aux_loss = None
+        if self.training:
+            aux_loss = self._compute_load_balancing_loss(router_probs, expert_indices)
+        return output, aux_loss
+
+    def _compute_load_balancing_loss(self, router_probs: torch.Tensor, expert_indices: torch.Tensor) -> torch.Tensor:
+        expert_mask = F.one_hot(expert_indices, num_classes=self.num_experts).float()
+        tokens_per_expert = expert_mask.sum(dim=[0, 1, 2]) / expert_mask.sum()
+        router_prob_mean = router_probs.mean(dim=[0, 1])
+        aux_loss = torch.sum(tokens_per_expert * router_prob_mean) * self.num_experts
+        return aux_loss * self.load_balancing_weight
+
+
+class MoETransformerBlock(nn.Module):
+    def __init__(
+        self,
+        d_model: int,
+        n_heads: int,
+        d_ff: int,
+        max_seq_len: int,
+        num_experts: int = 8,
+        top_k: int = 2,
+        dropout: float = 0.1,
+        use_moe: bool = True,
+    ):
+        super().__init__()
+        self.use_moe = use_moe
+        self.attention = MultiHeadAttention(d_model, n_heads, max_seq_len, dropout)
+        if use_moe:
+            self.feed_forward = MixtureOfExperts(d_model, d_ff, num_experts, top_k, dropout)
+        else:
+            self.feed_forward = FeedForward(d_model, d_ff, dropout)
+        self.norm1 = nn.RMSNorm(d_model)
+        self.norm2 = nn.RMSNorm(d_model)
+        self.dropout = nn.Dropout(dropout)
+
+    def forward(self, x):
+        attn_out = self.attention(self.norm1(x))
+        x = x + self.dropout(attn_out)
+        if self.use_moe:
+            ff_out, aux_loss = self.feed_forward(self.norm2(x))
+            x = x + self.dropout(ff_out)
+            return x, aux_loss
+        else:
+            ff_out = self.feed_forward(self.norm2(x))
+            x = x + self.dropout(ff_out)
+            return x, None
+
+
+class MinimalLLM(nn.Module):
+    def __init__(self, config: ModelConfig):
+        super().__init__()
+        self.config = config
+        self.token_embedding = nn.Embedding(config.vocab_size, config.d_model)
+        self.position_dropout = nn.Dropout(config.dropout)
+        self.transformer_blocks = nn.ModuleList([
+            TransformerBlock(config.d_model, config.n_heads, config.d_ff, config.max_seq_len, config.dropout)
+            for _ in range(config.n_layers)
+        ])
+        self.norm = nn.RMSNorm(config.d_model)
+        self.output_dropout = nn.Dropout(config.dropout)
+        self.lm_head = nn.Linear(config.d_model, config.vocab_size, bias=False)
+        self.lm_head.weight = self.token_embedding.weight
+        self.apply(self._init_weights)
+
+    def _init_weights(self, module):
+        if isinstance(module, nn.Linear):
+            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
+            if module.bias is not None:
+                torch.nn.init.zeros_(module.bias)
+        elif isinstance(module, nn.Embedding):
+            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
+
+    def forward(self, x):
+        x = self.token_embedding(x) * math.sqrt(self.config.d_model)
+        x = self.position_dropout(x)
+        for block in self.transformer_blocks:
+            x = block(x)
+        x = self.norm(x)
+        x = self.output_dropout(x)
+        logits = self.lm_head(x)
+        return logits
+
+
+class MoEMinimalLLM(nn.Module):
+    def __init__(self, config: MoEModelConfig):
+        super().__init__()
+        self.config = config
+        self.token_embedding = nn.Embedding(config.vocab_size, config.d_model)
+        self.position_dropout = nn.Dropout(config.dropout)
+        self.transformer_blocks = nn.ModuleList([
+            MoETransformerBlock(
+                config.d_model,
+                config.n_heads,
+                config.d_ff,
+                config.max_seq_len,
+                config.num_experts,
+                config.expert_top_k,
+                config.dropout,
+                use_moe=config.should_use_moe(i),
+            )
+            for i in range(config.n_layers)
+        ])
+        self.norm = nn.RMSNorm(config.d_model)
+        self.output_dropout = nn.Dropout(config.dropout)
+        self.lm_head = nn.Linear(config.d_model, config.vocab_size, bias=False)
+        self.lm_head.weight = self.token_embedding.weight
+        self.apply(self._init_weights)
+
+    def _init_weights(self, module):
+        if isinstance(module, nn.Linear):
+            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
+            if module.bias is not None:
+                torch.nn.init.zeros_(module.bias)
+        elif isinstance(module, nn.Embedding):
+            torch.nn.init.normal_(module.weight, mean=0.0, std=0.02)
+
+    def forward(self, x, return_aux_loss=True):
+        x = self.token_embedding(x) * math.sqrt(self.config.d_model)
+        x = self.position_dropout(x)
+        aux_losses = []
+        for block in self.transformer_blocks:
+            x, aux_loss = block(x)
+            if aux_loss is not None and return_aux_loss:
+                aux_losses.append(aux_loss)
+        x = self.norm(x)
+        x = self.output_dropout(x)
+        logits = self.lm_head(x)
+        total_aux_loss = sum(aux_losses) if aux_losses else None
+        if return_aux_loss:
+            return logits, total_aux_loss
+        return logits
+
+
